chore(protocol): channel out coverage fixes
