+++ crates/protocol/src/channel_out.rs
@@ -147,50 +147,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{CompressorResult, CompressorWriter, SingleBatch, SpanBatch};
-    use alloy_primitives::Bytes;
-
-    #[derive(Debug, Clone, Default)]
-    struct MockCompressor {
-        pub compressed: Option<Bytes>,
-    }
-
-    impl CompressorWriter for MockCompressor {
-        fn write(&mut self, data: &[u8]) -> CompressorResult<usize> {
-            let data = data.to_vec();
-            let written = data.len();
-            self.compressed = Some(Bytes::from(data));
-            Ok(written)
-        }
-
-        fn flush(&mut self) -> CompressorResult<()> {
-            Ok(())
-        }
-
-        fn close(&mut self) -> CompressorResult<()> {
-            Ok(())
-        }
-
-        fn reset(&mut self) {
-            self.compressed = None;
-        }
-
-        fn len(&self) -> usize {
-            self.compressed.as_ref().map(|b| b.len()).unwrap_or(0)
-        }
-
-        fn read(&mut self, buf: &mut [u8]) -> CompressorResult<usize> {
-            let len = self.compressed.as_ref().map(|b| b.len()).unwrap_or(0);
-            buf[..len].copy_from_slice(self.compressed.as_ref().unwrap());
-            Ok(len)
-        }
-    }
-
-    impl ChannelCompressor for MockCompressor {
-        fn get_compressed(&self) -> Vec<u8> {
-            self.compressed.as_ref().unwrap().to_vec()
-        }
-    }
+    use crate::test_utils::MockCompressor;
+    use crate::{CompressorWriter, SingleBatch, SpanBatch};
 
     #[test]
     fn test_channel_out_reset() {
